user skin style setting is now working with new APIs
