+++ src/api/index.php
@@ -74,6 +74,7 @@ require __DIR__ . '/routes/email/mailchimp.php';
 require __DIR__ . '/routes/search.php';
 require __DIR__ . '/routes/users/users-admin.php';
 require __DIR__ . '/routes/users/user-current.php';
+require __DIR__ . '/routes/users/user-settings.php';
 
 // Run app
 $app->run();
